refactor: split archive.py into slug-aligned route modules (Seam B)
Extract character, world, course, settings, learner-profile, and diary
handlers from the monolithic archive router while preserving all 96 API
paths and the progress compat surface via include_router aggregation.

Closes #255
